+++ packages/component/src/page/trading/hooks/useExecutionReport.tsx
@@ -4,14 +4,10 @@ import { getOrderExecutionReportMsg } from "@/block/orders/getOrderExecutionRepo
 import { useSymbolsInfo, useWS, useEventEmitter } from "@orderly.network/hooks";
 
 export function useExecutionReport() {
-  const ws = useWS();
   const ee = useEventEmitter();
 
   const symbolsInfo = useSymbolsInfo();
   const symbolsInfoRef = useRef({});
-
-  const timer = useRef<any>({});
-  const timestamp = useRef<any>({});
 
   useEffect(() => {
     symbolsInfoRef.current = symbolsInfo;
@@ -35,37 +31,48 @@ export function useExecutionReport() {
       }
     };
 
-    const unsubscribe = ws.privateSubscribe(
-      {
-        id: "executionreport",
-        event: "subscribe",
-        topic: "executionreport",
-        ts: Date.now(),
-      },
-      {
-        onMessage: (data: any) => {
-          // console.log("useExecutionReport", data);
-          showToast(data);
+    const handler = (data: any) => {
+      // console.log("orders:changed", data);
+      showToast(data);
+    };
 
-          const key = `${data.status}_${data.orderId}_${data.timestamp}`;
+    ee.on("orders:changed", handler);
 
-          // 如果100豪秒内有一样的订单状态更新
-          if (
-            timestamp.current[key] &&
-            Date.now() - timestamp.current[key] < 100
-          ) {
-            timer.current[key] && clearTimeout(timer.current[key]);
-          }
-          timer.current[key] = setTimeout(() => {
-            showToast(data);
-            delete timer.current[key];
-            delete timestamp.current[key];
-            ee.emit("orders:changed", data);
-          }, 100);
-          timestamp.current[key] = Date.now();
-        },
-      }
-    );
-    return () => unsubscribe();
+    return () => {
+      ee.off("orders:changed", handler);
+    };
+
+    // const unsubscribe = ws.privateSubscribe(
+    //   {
+    //     id: "executionreport",
+    //     event: "subscribe",
+    //     topic: "executionreport",
+    //     ts: Date.now(),
+    //   },
+    //   {
+    //     onMessage: (data: any) => {
+    //       // console.log("useExecutionReport", data);
+    //       showToast(data);
+
+    //       const key = `${data.status}_${data.orderId}_${data.timestamp}`;
+
+    //       // 如果100豪秒内有一样的订单状态更新
+    //       if (
+    //         timestamp.current[key] &&
+    //         Date.now() - timestamp.current[key] < 100
+    //       ) {
+    //         timer.current[key] && clearTimeout(timer.current[key]);
+    //       }
+    //       timer.current[key] = setTimeout(() => {
+    //         showToast(data);
+    //         delete timer.current[key];
+    //         delete timestamp.current[key];
+    //         ee.emit("orders:changed", data);
+    //       }, 100);
+    //       timestamp.current[key] = Date.now();
+    //     },
+    //   }
+    // );
+    // return () => unsubscribe();
   }, []);
 }