+++ packages/hooks/src/configProvider.tsx
@@ -17,6 +17,7 @@ import {
 import useConstant from "use-constant";
 import { NetworkId } from "@orderly.network/types";
 import { usePreLoadData } from "./usePreloadData";
+import { DataCenterProvider } from "./dataProvider";
 
 type RequireOnlyOne<T, U extends keyof T = keyof T> = Omit<T, U> &
   {
@@ -102,7 +103,7 @@ export const OrderlyConfigProvider = (
         // apiBaseUrl,
       }}
     >
-      {props.children}
+      <DataCenterProvider>{props.children}</DataCenterProvider>
     </OrderlyProvider>
   );
 };